requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
Pillow==7.0.0
pyTelegramBotAPI==3.7.3
requests==2.24.0
six==1.15.0
telebot==0.0.3
urllib3==1.25.10
